Basic flow established
The swagger client makes a request to an endpoint in your API.
The controller handles the request. It uses the repository to interact with the database.
The repository fetches or modifies the data in the database and returns the data to the controller.
The controller uses the DTO conversions to convert the data entities to DTOs.
The controller then sends the DTOs as a response to the client.